+++ tbot/machine/lab_noenv.py
@@ -6,6 +6,7 @@ import typing
 import pathlib
 import paramiko
 import tbot
+import tbot.config
 from . import machine
 
 
@@ -51,7 +52,7 @@ class MachineLabNoEnv(machine.Machine):
     @property
     def workdir(self) -> pathlib.PurePosixPath:
         if self._workdir is None:
-            raise Exception("No workdir specified")
+            raise tbot.config.InvalidConfigException("No workdir specified")
         return self._workdir
 
     @property