+++ tbot/main.py
@@ -17,8 +17,8 @@ def main() -> None:  # noqa: C901
     parser.add_argument(
         "testcase",
         type=str,
-        nargs="*",
-        default=None,
+        action="append",
+        default=[],
         help='name of the testcase to run (default: "uboot_checkout_and_build")',
     )
 
@@ -269,7 +269,7 @@ LOG:   "{logfile}\"""",
                     if tc in testcases:
                         tb.call(tc, **params)
                     else:
-                        raise Exception("Testcase not found")
+                        raise Exception(f"Testcase {tc!r} not found")
             else:
 
                 @tb.call
@@ -277,6 +277,24 @@ LOG:   "{logfile}\"""",
                     """ Default testcase is building U-Boot """
                     tb.call("uboot_checkout_and_build", **params)
 
+        except tbot.TestcaseFailure as f:
+            tbot.log.message(traceback.format_exc(), tbot.log.Verbosity.DEBUG)
+            tbot.log.message(
+                f"{tbot.log.has_color('31')}Testcase Failure{tbot.log.has_color('0')}: {f}",
+                tbot.log.Verbosity.ERROR,
+            )
+        except tbot.InvalidUsageException as e:
+            tbot.log.message(traceback.format_exc(), tbot.log.Verbosity.ERROR)
+            e_str = (
+                f" > {tbot.log.has_color('1')}"
+                + f"\n > {tbot.log.has_color('1')}".join(str(e).split("\n"))
+            )
+            tbot.log.message(
+                f"""\
+{tbot.log.has_color('31;1')}A programmer made a mistake{tbot.log.has_color('0')}:
+{e_str}""",
+                tbot.log.Verbosity.ERROR,
+            )
         except Exception:  # pylint: disable=broad-except
             tbot.log.message(traceback.format_exc(), tbot.log.Verbosity.ERROR)
         except KeyboardInterrupt: